refactor(slack): isolate message parsing
